refactor(gui): optimized layout calculations and modernized functions
- Optimized coordinate arithmetic, layout calculations, and coverage LUT generation using `f32::mul_add`.
- Converted various helper methods, constructors, and initialization functions to `const fn`.
- Refactored distance and timing checks to use squared arithmetic and `is_multiple_of`.
- Updated internal visibility of mux types, constants, and functions.

+++ crates/gui/src/macos.rs
@@ -1,5 +1,5 @@
 //! macOS window polish for the borderless, transparent shell: a compositor
-//! shadow shaped by the window's alpha, and WindowServer blur behind the
+//! shadow shaped by the window's alpha, and `WindowServer` blur behind the
 //! translucent regions.
 //!
 //! Blur uses `CGSSetWindowBackgroundBlurRadius` — the classic terminal
@@ -41,7 +41,7 @@ pub fn polish(window: &Window) {
 			return;
 		};
 		ns_window.setHasShadow(true);
-		let number = ns_window.windowNumber() as isize;
+		let number = ns_window.windowNumber();
 		CGSSetWindowBackgroundBlurRadius(CGSMainConnectionID(), number, 24);
 	}
 }
